requirements.txt
fastmcp>=0.1.0
snowflake-connector-python>=3.0.0
pydantic>=2.0.0
requests>=2.31.0
pyyaml>=6.0
typing-extensions>=4.0.0